+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.28;
-const unsigned char AutoSDKVersionString[] = "1.28.0";
+double AutoSDKVersionNumber = 1.29;
+const unsigned char AutoSDKVersionString[] = "1.29.0";